requirements.txt
aiobotocore==2.5.0
aiohttp==3.8.4
aioitertools==0.11.0
aiosignal==1.3.1
async-timeout==4.0.2
attrs==23.1.0
blinker==1.6.2
botocore==1.29.76
category-encoders==2.6.1
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
distlib==0.3.6
et-xmlfile==1.1.0
filelock==3.12.2
Flask==2.3.2
frozenlist==1.3.3
fsspec==2023.6.0
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
jmespath==1.0.1
joblib==1.2.0
MarkupSafe==2.1.3
multidict==6.0.4
numpy==1.25.0
openpy==0.0.2
openpyxl==3.1.2
packaging==23.1
pandas==2.0.2
patsy==0.5.3
pipenv==2023.6.18
platformdirs==3.8.0
python-dateutil==2.8.2
pytz==2023.3
requests==2.31.0
s3fs==2023.6.0
scikit-learn==1.2.2
scipy==1.10.1
six==1.16.0
sklearn-pandas==2.2.0
statsmodels==0.14.0
threadpoolctl==3.1.0
tzdata==2023.3
urllib3==1.26.16
virtualenv==20.23.1
virtualenv-clone==0.5.7
Werkzeug==2.3.6
wrapt==1.15.0
yarl==1.9.2
